chore: hoist /cloudproviders/aws to the top level
