+++ cmd/webhook/main.go
@@ -17,8 +17,8 @@ package main
 import (
 	"github.com/aws/karpenter-core/pkg/cloudprovider"
 	"github.com/aws/karpenter-core/pkg/webhooks"
-	awscloudprovider "github.com/aws/karpenter/pkg/cloudproviders/aws/cloudprovider"
-	awscontext "github.com/aws/karpenter/pkg/cloudproviders/aws/context"
+	awscloudprovider "github.com/aws/karpenter/pkg/cloudprovider"
+	awscontext "github.com/aws/karpenter/pkg/context"
 )
 
 func main() {
