+++ pkg/cloudprovider/amifamily/al2.go
@@ -25,7 +25,7 @@ import (
 	"github.com/aws/karpenter-core/pkg/cloudprovider"
 	"github.com/aws/karpenter-core/pkg/utils/resources"
 	"github.com/aws/karpenter/pkg/apis/awsnodetemplate/v1alpha1"
-	"github.com/aws/karpenter/pkg/cloudproviders/aws/cloudprovider/amifamily/bootstrap"
+	"github.com/aws/karpenter/pkg/cloudprovider/amifamily/bootstrap"
 )
 
 type AL2 struct {
