+++ pkg/cloudprovider/amifamily/bootstrap/bootstrap.go
