+++ pkg/cloudprovider/amifamily/bootstrap/bottlerocket.go
