+++ pkg/cloudprovider/amifamily/bootstrap/bottlerocketsettings.go
