+++ pkg/cloudprovider/amifamily/bootstrap/custom.go
