+++ pkg/cloudprovider/amifamily/resolver.go
@@ -34,7 +34,7 @@ import (
 	"github.com/aws/karpenter-core/pkg/utils/pretty"
 	"github.com/aws/karpenter-core/pkg/utils/ptr"
 	"github.com/aws/karpenter/pkg/apis/awsnodetemplate/v1alpha1"
-	"github.com/aws/karpenter/pkg/cloudproviders/aws/cloudprovider/amifamily/bootstrap"
+	"github.com/aws/karpenter/pkg/cloudprovider/amifamily/bootstrap"
 )
 
 var DefaultEBS = v1alpha1.BlockDevice{
