+++ pkg/cloudprovider/amifamily/userData.go
