+++ pkg/cloudprovider/instance.go
@@ -38,8 +38,8 @@ import (
 	"github.com/aws/karpenter-core/pkg/apis/provisioning/v1alpha5"
 	"github.com/aws/karpenter-core/pkg/operator/injection"
 	"github.com/aws/karpenter-core/pkg/operator/options"
-	awserrors "github.com/aws/karpenter/pkg/cloudproviders/aws/errors"
-	"github.com/aws/karpenter/pkg/cloudproviders/aws/utils"
+	awserrors "github.com/aws/karpenter/pkg/errors"
+	"github.com/aws/karpenter/pkg/utils"
 
 	"github.com/aws/karpenter-core/pkg/cloudprovider"
 	"github.com/aws/karpenter-core/pkg/scheduling"
