+++ pkg/cloudprovider/instancetype.go
@@ -36,7 +36,7 @@ import (
 	"github.com/aws/karpenter-core/pkg/scheduling"
 	"github.com/aws/karpenter-core/pkg/utils/resources"
 	"github.com/aws/karpenter/pkg/apis/awsnodetemplate/v1alpha1"
-	"github.com/aws/karpenter/pkg/cloudproviders/aws/cloudprovider/amifamily"
+	"github.com/aws/karpenter/pkg/cloudprovider/amifamily"
 )
 
 const (
