+++ pkg/cloudprovider/launchtemplate.go
@@ -39,9 +39,9 @@ import (
 	"github.com/aws/karpenter-core/pkg/operator/injection"
 	"github.com/aws/karpenter-core/pkg/utils/pretty"
 	"github.com/aws/karpenter/pkg/apis/awsnodetemplate/v1alpha1"
-	"github.com/aws/karpenter/pkg/cloudproviders/aws/cloudprovider/amifamily"
-	awscontext "github.com/aws/karpenter/pkg/cloudproviders/aws/context"
-	awserrors "github.com/aws/karpenter/pkg/cloudproviders/aws/errors"
+	"github.com/aws/karpenter/pkg/cloudprovider/amifamily"
+	awscontext "github.com/aws/karpenter/pkg/context"
+	awserrors "github.com/aws/karpenter/pkg/errors"
 )
 
 const (
