+++ pkg/cloudprovider/launchtemplate_test.go
@@ -48,7 +48,7 @@ import (
 	"github.com/aws/karpenter-core/pkg/utils/ptr"
 	"github.com/aws/karpenter/pkg/apis/awsnodetemplate/v1alpha1"
 
-	"github.com/aws/karpenter/pkg/cloudproviders/aws/cloudprovider/amifamily/bootstrap"
+	"github.com/aws/karpenter/pkg/cloudprovider/amifamily/bootstrap"
 	awstest "github.com/aws/karpenter/pkg/test"
 )
