+++ pkg/cloudprovider/subnets.go
@@ -30,7 +30,7 @@ import (
 	"github.com/aws/karpenter-core/pkg/utils/functional"
 	"github.com/aws/karpenter-core/pkg/utils/pretty"
 	"github.com/aws/karpenter/pkg/apis/awsnodetemplate/v1alpha1"
-	awscontext "github.com/aws/karpenter/pkg/cloudproviders/aws/context"
+	awscontext "github.com/aws/karpenter/pkg/context"
 )
 
 type SubnetProvider struct {
