+++ pkg/cloudprovider/testdata/al2_userdata_input.golden
