+++ pkg/cloudprovider/testdata/al2_userdata_merged.golden
