+++ pkg/cloudprovider/testdata/al2_userdata_unmerged.golden
