+++ pkg/cloudprovider/testdata/br_userdata_input.golden
