+++ pkg/cloudprovider/testdata/br_userdata_merged.golden
