+++ pkg/cloudprovider/testdata/br_userdata_unmerged.golden
