+++ pkg/fake/atomic.go
