+++ pkg/fake/ssmapi.go
